+++ lib/runners/processor/haml_lint.rb
@@ -153,7 +153,7 @@ module Runners
       # @see https://github.com/sds/haml-lint/blob/v0.35.0/lib/haml_lint/cli.rb#L110
       # @see https://github.com/ged/sysexits/blob/v1.2.0/lib/sysexits.rb#L96
       unless [65, 0].include?(status.exitstatus)
-        return Results::Failure.new(guid: guid, message: "HAML-Lint raises an unexpected error", analyzer: analyzer)
+        return Results::Failure.new(guid: guid, analyzer: analyzer)
       end
 
       add_rubocop_warnings_if_exists(stderr)