+++ test/smokes/tslint/expectations.rb
@@ -149,14 +149,7 @@ s.add_test(
 s.add_test(
   "raise-deprecated",
   type: "failure",
-  message: <<~MESSAGE,
-    --type-check is deprecated. You only need --project to enable rules which need type information.
-    Error at range.ts:2:18: Property 'min' does not exist on type 'string'.
-    Error at range.ts:2:31: Property 'middle' does not exist on type 'string'.
-    Error at range.ts:2:47: Property 'middle' does not exist on type 'string'.
-    Error at range.ts:2:63: Property 'max' does not exist on type 'string'.
-    Error at range.ts:11:24: Argument of type '{ min: number; middle: number; max: number; }' is not assignable to parameter of type 'string'.
-  MESSAGE
+  message: "The analysis failed due to an unexpected error. See the analysis log for details.",
   analyzer: { name: "TSLint", version: default_version },
   warnings: [
     {